# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/pay/withdraw/nft": {
            "post": {
                "description": "Withdraws NFTs on Solana.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payments"
                ],
                "summary": "Withdraw NFT",
                "operationId": "payment-withdraw-nft",
                "parameters": [
                    {
                        "description": "MintAddresses of NFTs to withdraw.",
                        "name": "mintAddresses",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/pay/withdraw/sol": {
            "post": {
                "description": "Withdraws chips as SOL on Solana.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payments"
                ],
                "summary": "Withdraw SOL",
                "operationId": "payment-withdraw-sol",
                "parameters": [
                    {
                        "description": "The amount of chips to withdraw.",
                        "name": "usdAmount",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
    },
...
}